Moved `get_code_with_max_score` to utilities
