+++ tools/hardcoding.py
@@ -2,7 +2,14 @@ import re
 
 import pandas as pd
 
-from tools.utilities import create_data_structure, download_code, get_selected_labs, setup_logger, write_output_to_csv
+from tools.utilities import (
+    create_data_structure,
+    download_code,
+    get_code_with_max_score,
+    get_selected_labs,
+    setup_logger,
+    write_output_to_csv,
+)
 
 logger = setup_logger(__name__)  # DEBUGGING
 
@@ -11,29 +18,6 @@ use_standalone = False
 IF_STATEMENT_REGEX = r'if\s*\((.*)\)'
 LITERAL_VALUE_IN_COMP_REGEX = r'\w+\s*==\s*((?:[\"\'][^\"\']*[\"\'])|\d+)'
 VAR_NAME_IN_COMP_REGEX = r'(\w+)\s*==\s*(?:[\"\'][^\"\']*[\"\']|\d+)'
-
-
-def get_code_with_max_score(user_id: int, lab: float, submissions: dict) -> str:
-    """Returns the first highest-scoring code submission for a student for a lab.
-
-    The "first" highest-scoring submission means the oldest submission with the highest score.
-
-    Args:
-        user_id (int): Find the highest-scoring submission for the student with this ID.
-        lab (float): Find the highest-scoring submission for this lab, e.g. lab 3.12.
-        submissions (dict): All of the student's submissions for this lab.
-
-    Returns:
-        code (str): The code for the first highest-scoring submission.
-    """
-
-    max_score = 0
-    code = submissions[user_id][lab][-1].code  # Choose a default submission
-    for sub in submissions[user_id][lab]:
-        if sub.max_score > max_score:
-            max_score = sub.max_score
-            code = sub.code
-    return code
 
 
 def remove_quotes(s: str) -> str:
